+++ scripts/gen-api.ts
@@ -8,18 +8,17 @@ import {
   DocBlock,
   DocCodeSpan,
   DocFencedCode,
+  DocNode,
   DocParagraph,
   DocPlainText,
   DocSection
 } from '@microsoft/tsdoc'
 import { configure, renderFile } from 'eta'
-import { writeFileSync } from 'fs-extra'
+import { isUndefined } from 'fonction'
+import { outputFileSync } from 'fs-extra'
 import { join, resolve } from 'path'
 
 const apiModel = new ApiModel()
-const apiPackage = apiModel.loadPackage(
-  resolve(__dirname, '..', 'temp', 'fonction.api.json')
-)
 
 const getSummary = ({ nodes }: DocSection): string => {
   const _nodeSection = nodes[0]
@@ -48,12 +47,61 @@ const getExampleCode = (
   return customBlocks
     .map(({ content }) => {
       const _docFencedCode = content.nodes[1]
-      if (_docFencedCode instanceof DocFencedCode) {
+      if (isDocFencedCode(_docFencedCode)) {
         const { code, language } = _docFencedCode
         return [language, code] as [string, string]
       }
     })
     .filter((_) => !!_)
+}
+
+const isDocParagraph = (docNode: DocNode): docNode is DocParagraph =>
+  docNode instanceof DocParagraph
+
+const isDocPlainText = (docNode: DocNode): docNode is DocPlainText =>
+  docNode instanceof DocPlainText
+
+const isDocCodeSpan = (docNode: DocNode): docNode is DocCodeSpan =>
+  docNode instanceof DocCodeSpan
+const isDocFencedCode = (docNode: DocNode): docNode is DocFencedCode =>
+  docNode instanceof DocFencedCode
+
+const getRemarks = (docBlock: DocBlock | undefined) => {
+  if (isUndefined(docBlock)) return []
+
+  const remarksMatrix = docBlock.content.nodes.map((_node) =>
+    isDocParagraph(_node)
+      ? _node.nodes
+          .map((_node) => getDocPlainText(_node) || getDocCodeSpan(_node))
+          .filter((_) => !!_)
+      : []
+  )
+
+  return remarksMatrix.map((remarks) => remarks.join(''))
+}
+
+const getDocCodeSpan = (docNode: DocNode): string => {
+  if (isDocCodeSpan(docNode)) {
+    return `\`${docNode.code}\``
+  }
+
+  return ''
+}
+
+const getDocFencedCode = (docNode: DocNode): string => {
+  if (isDocFencedCode(docNode)) {
+    return `\`${docNode.code}\``
+  }
+
+  return ''
+}
+
+const getDocPlainText = (docNode: DocNode): string => {
+  if (isDocPlainText(docNode)) {
+    return docNode.text
+  }
+
+  return ''
 }
 
 interface MdVariables {
@@ -63,6 +111,7 @@ interface MdVariables {
   signature: string
   tagName: string
   deprecated: boolean
+  remarks: string[]
 }
 
 const render = async (mdArgs: MdVariables): Promise<string> => {
@@ -77,7 +126,7 @@ const render = async (mdArgs: MdVariables): Promise<string> => {
 }
 
 const generate = (path: string, content: string): void =>
-  writeFileSync(path, content)
+  outputFileSync(path, content)
 
 const mapMember = async (
   member: ApiItem
@@ -93,11 +142,13 @@ const mapMember = async (
       summarySection,
       customBlocks,
       modifierTagSet,
-      deprecatedBlock
+      deprecatedBlock,
+      remarksBlock
     } = member.tsdocComment
 
+    const remarks = getRemarks(remarksBlock)
     const isDeprecated = !!deprecatedBlock
-    const tagName = modifierTagSet.nodes[0].tagName
+    const tagName = modifierTagSet.nodes[0]?.tagName ?? ''
     const description = getSummary(summarySection)
     const examples = getExampleCode(customBlocks)
     const md = await render({
@@ -106,7 +157,8 @@ const mapMember = async (
       examples,
       signature,
       tagName,
-      deprecated: isDeprecated
+      deprecated: isDeprecated,
+      remarks
     })
 
     return {
@@ -115,7 +167,17 @@ const mapMember = async (
     }
   }
 }
-const run = async () => {
+const run = async ({
+  version,
+  apiJsonPath,
+  editLink = true
+}: {
+  version: string
+  apiJsonPath: string
+  editLink?: boolean
+}): Promise<void> => {
+  const apiPackage = apiModel.loadPackage(apiJsonPath)
+
   apiPackage.members.forEach(async (root) => {
     const functionContents = await Promise.all(
       root.members.filter(({ kind }) => kind === 'Variable').map(mapMember)
@@ -130,9 +192,15 @@ const run = async () => {
     const merged = mdFunctions.join('\n')
     const mergedTypesMd = mdTypes.join('\n')
 
+    const frontmatter = `---
+editLink: false
+---
+
+`
+
     generate(
-      resolve(__dirname, '..', 'docs', 'api', 'index.md'),
-      `# API
+      resolve(__dirname, '..', 'docs', 'api', version, 'index.md'),
+      `${editLink ? '' : frontmatter}# API
 
 ## Functions
 
@@ -146,4 +214,11 @@ ${mergedTypesMd}
   })
 }
 
-run()
+if (require.main === module) {
+  run({
+    version: '',
+    apiJsonPath: resolve(__dirname, '..', 'temp', 'fonction.api.json')
+  })
+}
+
+export { run }